cmds: Clean udp_test
